feat: use probe to get metadata and update record

+++ monorepo/internal/http/dtos/response/video_response.go
@@ -1,8 +1,8 @@
 package response
 
 type VideoResponse struct {
-	VideoID  string `json:"video_id"`
-	Title    string `json:"title"`
-	Duration int    `json:"duration"`
-	Status   string `json:"status"`
+	VideoID  string  `json:"video_id"`
+	Title    string  `json:"title"`
+	Duration float64 `json:"duration"`
+	Status   string  `json:"status"`
 }
